+++ src/lib/Components/Inbox/Conversations/Message.tsx
@@ -4,6 +4,9 @@ import * as Relay from "react-relay"
 
 import { BodyText, MetadataText, SmallHeadline } from "../Typography"
 
+import ImagePreview from "./Previews/ImagePreview"
+import PDFPreview from "./Previews/PDFPreview"
+
 import styled from "styled-components/native"
 import colors from "../../../../data/colors"
 
@@ -53,15 +56,26 @@ const ImagePreviewContainer = styled.View`
   marginBottom: 10
 `
 
+const PDFPreviewContainer = styled.View`
+  marginBottom: 10
+`
+
 interface Props extends RelayProps {
   senderName: string
   artworkPreview?: JSX.Element
-  imagePreview?: JSX.Element
 }
 
 export class Message extends React.Component<Props, any> {
   render() {
     const date = moment(this.props.message.created_at).fromNow(true)
+    const attachments = this.props.message.attachments
+    let hasImageAttachment
+    let hasPDFAttachment
+    if (attachments.length > 0) {
+      hasImageAttachment = attachments[0].content_type === "image/jpeg"
+      hasPDFAttachment = attachments[0].content_type === "application/pdf"
+    }
+
     return (
       <Container>
         <Avatar />
@@ -71,7 +85,14 @@ export class Message extends React.Component<Props, any> {
             <MetadataText>{date}</MetadataText>
           </Header>
           {this.props.artworkPreview && <ArtworkPreviewContainer>{this.props.artworkPreview}</ArtworkPreviewContainer>}
-          {this.props.imagePreview && <ImagePreviewContainer>{this.props.imagePreview}</ImagePreviewContainer>}
+          {hasImageAttachment &&
+            <ImagePreviewContainer>
+              <ImagePreview imageAttachment={attachments[0]} />
+            </ImagePreviewContainer>}
+          {hasPDFAttachment &&
+            <PDFPreviewContainer>
+              <PDFPreview pdfAttachment={attachments[0]} />
+            </PDFPreviewContainer>}
           <BodyText>{this.props.message.raw_text.split("\n\nAbout")[0]}</BodyText>
         </TextContainer>
       </Container>
@@ -86,6 +107,11 @@ export default Relay.createContainer(Message, {
         raw_text
         created_at
         is_from_user
+        attachments {
+          content_type
+          download_url
+          file_name
+        }
       }
     `,
   },
@@ -96,5 +122,12 @@ interface RelayProps {
     raw_text: string | null
     created_at: string | null
     is_from_user: boolean
+    attachments: Array<
+      {
+        content_type: string
+        download_url: string
+        file_name: string
+      }
+    >
   }
 }